requirements.txt
nicegui>=2.15.0
yt-dlp>=2025.3.31
ffmpeg-python>=0.2.0
openai>=1.75.0
pydantic>=2.11.0
click>=8.1.8 